Move `text_generation.py`
This move makes it discoverable by MkDocs and simplifies repo structure.
