+++ docs/api/services.md
@@ -1,11 +1,4 @@
 # Services
 
-<!-- ### :::app.services.nlp.text_generation.TextGenerationConfig
-### :::app.services.nlp.text_generation.TextGenerationParameters -->
-
-<!-- ::: app.services.nlp.text_generation.TextGenerationConfig
-    :docstring:
-    :members: -->
-
 ::: app.services.base
 ::: app.services.text_generation
