+++ src/app/services/text_generation.py
